arch: arm: rename expection header
Rename exception header and use the same name as all architecture ports.
