+++ arch/arm/include/cortex_m/exc.h
@@ -23,7 +23,7 @@
 #else
 
 #include <cmsis_core.h>
-#include <zephyr/arch/arm/exc.h>
+#include <zephyr/arch/arm/exception.h>
 #include <zephyr/irq_offload.h>
 
 #ifdef __cplusplus
