+++ include/zephyr/arch/arm/arch.h
@@ -23,7 +23,7 @@
 #define sys_define_gpr_with_alias(name1, name2) union { uint32_t name1, name2; }
 
 #include <zephyr/arch/arm/thread.h>
-#include <zephyr/arch/arm/exc.h>
+#include <zephyr/arch/arm/exception.h>
 #include <zephyr/arch/arm/irq.h>
 #include <zephyr/arch/arm/error.h>
 #include <zephyr/arch/arm/misc.h>
