+++ include/zephyr/arch/arm/asm_inline_gcc.h
@@ -20,7 +20,7 @@
 
 #include <zephyr/toolchain.h>
 #include <zephyr/types.h>
-#include <zephyr/arch/arm/exc.h>
+#include <zephyr/arch/arm/exception.h>
 
 #if defined(CONFIG_CPU_AARCH32_CORTEX_R) || defined(CONFIG_CPU_AARCH32_CORTEX_A)
 #include <zephyr/arch/arm/cortex_a_r/cpu.h>
